requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn>=20.1.0
psycopg2-binary>=2.9.0
